IMPALA-7144: Re-enable TestDescribeTableResults
This patch makes the TestDescribeTableResults more robust by only
comparing the information that the authorization cares about instead
of comparing all output in DESCRIBE. This change will avoid any unnecessary
changes to AuthorizationTest if HMS updates the DESCRIBE output.

The test is also updated to support standalone execution without relying
on other tests be executed first since it can cause the test to be flaky
especially if the tests in AuthorizationTest are executed in parallel.

Testing:
- Ran all FE tests

Cherry-picks: not for 2.x

Change-Id: I3aeaecf5b6d906a66d338e165a6d506e3964563f
Reviewed-on: http://gerrit.cloudera.org:8080/10643
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>

Conflicts:
	fe/src/test/java/org/apache/impala/analysis/AuthorizationTest.java