IMPALA-6969: add AC last queued reason to profile
The reason is updated during initial admission and when the query is at
the head of the queue but can't be admitted. It is not updated while
the query is in the middle of the queue.

Together with the async admission change, this makes it possible to
determine from the profile why the query has not been admitted yet.

Testing:
Added admission control tests that check that the
string is set for queries queued based both on the
query count and the max memory.

Looped the tests overnight to confirm non-flakiness.

Change-Id: Ida9b75dc50dfb7a27f59deda91bad6ac838130a1
Reviewed-on: http://gerrit.cloudera.org:8080/10731
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>