IMPALA-8516. Update maven for Jenkins builds
This changes Maven to download and install on both Ubuntu and Redhat for
the Jenkins builds (previously it was only installed on Redhat).

The version number is kept at 3.5.4 even though a newer release is
available upstream. The new release fails to build Impala due to an
XML-parsing bug causing it to fail to resolve the parquet pom [1]

This should hopefully address some of the hang issues we've seen
previously with the older version of Maven that shipped with the version
of Ubuntu we have on Ubuntu 16.04.

[1] https://github.com/codehaus-plexus/plexus-utils/issues/65

Change-Id: I793409eb4e9f4533b75bfe089a497c0ea62ad1ff
Reviewed-on: http://gerrit.cloudera.org:8080/13268
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>