IMPALA-10579: Fix usage of RemoteIterator in FileSystemUtil
HDFS FileSystem provides a listStatusIterator() API for listing remote
storage using a RemoteIterator. We use it to list files when loading
table file metadata.

It's not guaranteed that a RemoteIterator can survive when its hasNext()
or next() throws IOExceptions. We should stop the loop in this case.
Otherwise, we may go into a dead loop.

Without HADOOP-16685, it's also not guaranteed that
FileSystem.listStatusIterator() will thrown a FileNotFoundException when
the path doesn't exist.

This patch refactors the file listing iterators so we don't need to
depend on these two assumptions.

Tests:
 - Run test_insert_stress.py. Verified the non-existing subdirs are
   skipped.

Change-Id: I859bd4f976c51a34eb6a03cefd2ddcdf11656cea